backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.10

# Pydantic (latest with Python 3.13 prebuilt wheels)
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator

# Authentication
pyjwt[crypto]>=2.9.0
passlib[bcrypt]>=1.7.4

# AI and External APIs
google-generativeai>=0.8.0
anthropic>=0.40.0
httpx>=0.27.0
PyGithub>=2.5.0

# Document processing
python-docx>=1.1.2
markdown>=3.7
weasyprint>=63.0

# Utilities
gitpython>=3.1.43
aiofiles>=24.1.0
python-magic>=0.4.27
